placer-service/pyproject.toml
[tool.poetry]
name = "fly-placer"
version = "0.1.0"
description = "An auto-placer for Fly.io deployments based on real-time traffic patterns."
authors = ["Your Name <you@example.com>"]
license = "MIT"
packages = [
    { include = "automation" },
    { include = "utils" },
    { include = "monitoring" },
    { include = "prediction" }
]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
requests = "^2.27.1"
python-dotenv = "^0.20.0"
python-dateutil = "^2.9.0.post0"
pyyaml = "^6.0.2"
fastapi = {extras = ["standard"], version = "^0.115.0"}
fastapi-limiter = "^0.1.6"
numpy = "^2.1.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
flake8 = "^4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]